requirements.txt
click==8.1.7
pyyaml==6.0.1
gitpython==3.1.41
colorama==0.4.6
